change float to double by the card entity

+++ backend/src/main/java/com/github/ulrikewerner/backend/entities/Card.java
@@ -3,6 +3,6 @@ package com.github.ulrikewerner.backend.entities;
 public record Card(
         String id,
         String team,
-        float pointsPerGame
+        double pointsPerGame
 ) {
 }
